feat: :rocket: functional API consumption
The web page based on the Space X missions is created, which contains a "card" of each one of them, modals with additional information and the video of each mission. all the modals are brought from the api and are inserted when loading the page along with the "Cards" of each mission, the videos of each modal make the page load slow. It must be corrected so that the video is only inserted in the modal when it is displayed.